HowToUseJpaのkotlin変換3

+++ src/main/kotlin/com/example/repository/HTURepository.kt
@@ -4,6 +4,7 @@ import com.example.model.HTUData
 import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor
 import org.springframework.stereotype.Repository
+//import samples.model.HTUData
 
 
 /**
@@ -11,7 +12,7 @@ import org.springframework.stereotype.Repository
  */
 
 @Repository
-interface HTURepository : JpaRepository<HTUData, Long> , JpaSpecificationExecutor<HTUData> {
+interface HTURepository : JpaRepository<HTUData, Int> , JpaSpecificationExecutor<HTUData> {
 
     fun findByid(id: Int?): HTUData
